requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
astropy==6.0.0
astropy-iers-data==0.2024.2.19.0.28.47
beautifulsoup4==4.12.2
billiard==4.2.0
black==25.11.0
boto3==1.34.40
botocore==1.34.40
celery==5.3.6
celery-progress==0.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cryptography==42.0.5
distlib==0.3.8
Django==4.2.16
django-anymail==10.2
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-composer==0.0.5
django-compressor==4.4
django-debug-toolbar==5.1.0
django-health-check==3.18.1
django-jazzmin==2.6.1
django-leaflet==0.30.0
django-libsass==0.9
django-ninja==1.3.0
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-storages==1.14.3
django-upgrade==1.20.0
djangorestframework==3.14.0
dnspython==2.8.0
email-validator==2.3.0
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.33
idna==3.6
importlib-resources==5.13.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.5
libsass==0.22.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.5.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pyerfa==2.0.1.1
pytest==8.3.5
pytest-benchmark==5.1.0
pytest-cov==5.0.0
pytest-django==4.7.0
pytest-mock==3.12.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-mock==1.11.0
rjsmin==1.2.1
ruff==0.14.5
s3transfer==0.10.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tokenize-rt==6.0.0
types-jmespath==1.0.2.20240106
types-requests==2.31.0.20240311
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
whitenoise==6.7.0
